nexus_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
billiard==4.2.1
bleach==6.3.0
boto3==1.39.4
botocore==1.39.4
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.5
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.1
django-appconf==1.1.0
django-celery-beat==2.8.1
django-compressor==4.5.1
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-mysql==4.17.0
django-rosetta==0.10.2
django-storages==1.14.6
django-timezone-field==7.1
django-weasyprint==2.4.0
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
ecdsa==0.19.1
et_xmlfile==2.0.0
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.58.5
frozenlist==1.7.0


gunicorn==23.0.0
h3==4.3.1
html5lib==1.1
humanize==4.13.0
idna==3.10
jmespath==1.0.1
joblib==1.5.2
kombu==5.5.4
lxml==6.0.0
Markdown==3.8
MarkupSafe==3.0.2
multidict==6.6.3
nltk==3.9.2
numpy==2.3.1
openlocationcode==1.0.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.0
pdf2image>=1.17.0
phonenumbers==9.0.9
pillow==11.3.0
polib==1.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMySQL==1.1.2
pypdf==5.7.0
pyphen==0.17.2
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.2
redis==6.4.0
regex==2025.10.23
reportlab==4.4.2
rjsmin==1.2.2
s3transfer==0.13.0
scipy==1.16.2
sendgrid==6.12.4
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
svglib==1.5.1
timezonefinder==8.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
tqdm==4.67.1
twilio==9.6.4
# Email via Microsoft Graph (Anymail)
django-anymail[msgraph]==12.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
valkey==6.1.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
xhtml2pdf==0.2.17
yarl==1.20.1
zopfli==0.2.3.post1
